Create TooltipContainer component
