+++ src/components/LineChart/utilities/index.ts
@@ -1,2 +1,1 @@
-export {eventPoint} from './event-point';
 export {yAxisMinMax} from './y-axis-min-max';
