+++ src/components/TooltipContainer/index.ts
@@ -1,0 +1,1 @@
+export {TooltipContainer} from './TooltipContainer';
