+++ src/components/index.ts
@@ -2,3 +2,4 @@ export {LineChart} from './LineChart';
 export {NormalizedStackedBar} from './NormalizedStackedBar';
 export {Sparkline} from './Sparkline';
 export {YAxis} from './YAxis';
+export {TooltipContainer} from './TooltipContainer';
